feat: allow safe area configuration for bottom sheets (#2)
Allows the `ignoreSafeArea` param for the showBottomSheet funciton.
This setting defaults to "true", such that any `SafeArea` widget
is ignored in a bottom sheet (barring the bottom padding value).
When set to "false", a safe area widget is respected and the
bottomsheet padded accordingly. This is most likely used
with "fullscreen" bottom sheets.