feat: allow barrier dismiss in dialogs (#3)
This add the builder option and possibility to allow
dialog dismissal via backdrop press. Be aware that
Scaffold and Material App are full screen by default.
So to use a "small modal" of some sort, you need to
provide theme data or use "Material" widgets to provide
defaults. (e.g. in the example app)