feat: use native navigation instead of getX navigation (#6)
This removes get (getX) as the underlying navigation engine.
Since GetX also routes with the native navigation stack,
this can be directly used instead of passed to get which
then in turn passes it to flutter. While doing this, the services
were redesigned and all navigation related matters are now in
the navigation service only.

BREAKING CHANGE: The `DialogService` was deleted.
Dialog related methods are now in the `NavigationService`.

BREAKING CHANGE: The `BottomSheetService` was deleted.
All related methods are now in the `NavigationService`.

BREAKING CHANGE: the close methods for dialog and
sheet (`closeDialog` and `closeSheet`) are removed. They
are unified in the method `closeOverlay` in the `NavigationService`.

BREAKING CHANGE: Since getX is removed, the base logic of
the entry point in the app could be simplified. Instead of
the key and observer being methods, they are now static
initialized fields on the navigation service.

BRAKING CHANGE: The `preventDuplicates` parameter for the
navigation method is gone. Currently, with native
navigation, the user of the package is responsible to
have an overview of the navigation stack.